+++ src/rpc/config.ts
@@ -1,5 +1,5 @@
 import configstore from '../common/configstore';
-import { BlockchainNodes } from '../types/node';
+import { BlockchainNodes } from '../types';
 
 export const baseUrl = (): string => {
   let url = 'https://privatebeta.henesis.io';
@@ -15,7 +15,7 @@ export const baseUrl = (): string => {
   return url;
 };
 export const rpcVersion = 'v1';
-export const trustedNodeBaseUrl = 'http://network.henesis.io';
+export const trustedNodeBaseUrl = 'http://tn.henesis.io';
 export const blockchainKinds = new BlockchainNodes(
   ['mainnet', 'ropsten', 'rinkeby'], //ethereum
   ['mainnet', 'baobab'], //klaytn